cmitt integer

+++ database/migrations/2021_08_12_222447_create_permission_tables.php
@@ -36,7 +36,7 @@ class CreatePermissionTables extends Migration
             $table->string('guard_name'); // For MySQL 8.0 use string('guard_name', 125);
             $table->timestamps();
 
-            $table->unsignedBigInteger('idCompany');
+            $table->unsignedInteger('idCompany');
             $table->foreign('idCompany')->references('id')->on('empresa');
 
             $table->unique(['name', 'guard_name', 'idCompany']);
